+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,7 +1,7 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '1.8.1-dev'.freeze
-    EXTENSION_BUILD = '201911191906'.freeze
+    EXTENSION_VERSION = '1.8.3-dev'.freeze
+    EXTENSION_BUILD = '201912121537'.freeze
 
 end